microservices/contentvideo-service/requirements.txt
psycopg2-binary
python-dotenv
django
boto3